refactor(zip): Clean up some small issues in the generator.
